Move `TransactionReceipt.kt` file and update imports
