+++ lib/src/main/kotlin/com/swmansion/starknet/data/serializers/TransactionPayloadSerializer.kt
@@ -1,6 +1,7 @@
 package com.swmansion.starknet.data.serializers
 
 import com.swmansion.starknet.data.types.Felt
+import com.swmansion.starknet.data.types.TransactionType
 import com.swmansion.starknet.data.types.transactions.*
 import com.swmansion.starknet.extensions.add
 import com.swmansion.starknet.extensions.put
