+++ lib/src/main/kotlin/com/swmansion/starknet/data/serializers/TransactionReceiptPolymorphicSerializer.kt
@@ -1,6 +1,6 @@
 package com.swmansion.starknet.data.serializers
 
-import com.swmansion.starknet.data.types.transactions.*
+import com.swmansion.starknet.data.types.*
 import kotlinx.serialization.DeserializationStrategy
 import kotlinx.serialization.SerializationException
 import kotlinx.serialization.json.*
