+++ lib/src/main/kotlin/com/swmansion/starknet/data/serializers/TransactionTracePolymorphicSerializer.kt
@@ -1,5 +1,6 @@
 package com.swmansion.starknet.data.serializers
 
+import com.swmansion.starknet.data.types.TransactionType
 import com.swmansion.starknet.data.types.transactions.*
 import kotlinx.serialization.DeserializationStrategy
 import kotlinx.serialization.SerializationException
