+++ lib/src/main/kotlin/com/swmansion/starknet/signer/Signer.kt
@@ -2,7 +2,6 @@ package com.swmansion.starknet.signer
 
 import com.swmansion.starknet.data.TypedData
 import com.swmansion.starknet.data.types.*
-import com.swmansion.starknet.data.types.transactions.*
 
 /**
  * Transaction signer.
