+++ lib/src/main/kotlin/com/swmansion/starknet/signer/StarkCurveSigner.kt
@@ -3,7 +3,6 @@ package com.swmansion.starknet.signer
 import com.swmansion.starknet.crypto.StarknetCurve
 import com.swmansion.starknet.data.TypedData
 import com.swmansion.starknet.data.types.*
-import com.swmansion.starknet.data.types.transactions.*
 
 /**
  * Signer implementing a stark curve signature (default signature used on Starknet).
