+++ lib/src/test/kotlin/starknet/data/TransactionStatusMappingTest.kt
@@ -1,6 +1,6 @@
 package starknet.data
 
-import com.swmansion.starknet.data.types.transactions.TransactionStatus
+import com.swmansion.starknet.data.types.TransactionStatus
 import kotlinx.serialization.SerializationException
 import kotlinx.serialization.json.Json
 import org.junit.jupiter.api.Assertions.assertEquals
